refactor: move tokenizer utilities into package
