+++ scripts/chat_template_checker.py
@@ -70,13 +70,13 @@ from scripts.chat_template_mock_trajectories import (  # noqa: E402
     SingleTurnTrajectory,
     ToolAgentTrajectory,
 )
-from verl.utils.chat_template import apply_chat_template  # noqa: E402
-from verl.utils.continuous_token_wiring import (  # noqa: E402
+from verl.utils.tokenizer import normalize_token_ids  # noqa: E402
+from verl.utils.tokenizer.chat_template import apply_chat_template  # noqa: E402
+from verl.utils.tokenizer.continuous_token_wiring import (  # noqa: E402
     CONTINUOUS_TOKEN_BUILDER_FAMILIES,
     create_continuous_token_builder,
     resolve_continuous_token_model_family,
 )
-from verl.utils.tokenizer import normalize_token_ids  # noqa: E402
 
 CheckLayer = Literal["raw-template", "continuous-token"]
