+++ verl/utils/tokenizer/__init__.py
@@ -1,0 +1,29 @@
+# Copyright 2024 Bytedance Ltd. and/or its affiliates
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from .tokenizer import (
+    build_multimodal_processor_inputs,
+    get_processor_token_id,
+    hf_processor,
+    hf_tokenizer,
+    normalize_token_ids,
+)
+
+__all__ = [
+    "hf_tokenizer",
+    "hf_processor",
+    "normalize_token_ids",
+    "build_multimodal_processor_inputs",
+    "get_processor_token_id",
+]
